Move new utility components to subfolder
