+++ web/tsconfig.json
@@ -17,6 +17,6 @@
     "jsx": "react",
     "downlevelIteration": true
   },
-    "include": ["src"],
-    "exclude" : ["node_modules", "build"]
+  "include": ["src"],
+  "exclude": ["node_modules", "build"]
 }
